pyproject.toml
[project]
name = "pillow_degas"
description = "Atari DEGAS image loader for Pillow"
version = "0.2.0"
authors = [
    { name = "Gareth Davidson", email = "gaz@bitplane.net" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "pillow"
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "coverage",
    "pytest-cov",
    "build",
    "twine",
    "ruff",
    "mkdocs",
    "mkdocs-material",
    "pydoc-markdown"
]

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true
